# users_ms/src/main/resources/static/swagger.yaml
openapi: 3.0.0
info:
  title: Nimbus - Usuários
  description: >-
    Micro-serviço responsável pelo gerenciamento de dados relacionados aos
    usuários do sistema Nimbus
  version: 1.0.5
servers:
  - url: http://localhost:3000
    description: Servidor local porta 3000
paths:
...
  /athletes/ms-get-athlete/{id}:
    get:
      tags:
        - Athletes
      summary: Retorna as informações do atleta.
      description: Retorna informações de persona do atleta por seu ID.
      parameters:
        - in: path
          name: Id
          schema:
            type: string
          required: true
          description: Id(UUID) do usuário.
      responses:
        200:
          description: Atleta encontrado.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/ResponseMessage"
              example:
                clientMessage: null
                serverMessage: null
                data:
                  firstName: Melissa
                  lastName: Cavalcante
                  birthDate: 2004-10-07
                  phone: 84244562331
                  picture: null
                  category: sub-20
                  isStarting: true
        404:
          description: Usuário não encontrado.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/ResponseMessage"
              example:
                clientMessage: Informações de usuário não encontradas
                serverMessage: null
                data: null
  /athletes/generate-csv:
    get:
      tags:
        - Athletes
      summary: Cria o arquivo csv.         
      description: Cria o arquivo csv com as informações dos atletas.
      parameters:
        - in: path
          name: Ids
          schema:
            type: string
          required: true
          description: List Id (UUID) do usuário.
      responses:
        200:
            description: criação do arquivo feita com sucesso.
            content:
              application/json:
                schema:
                  $ref: "#components/schemas/ResponseMessage"
                example:
                  clientMessage: null
                  serverMessage: null
                  data: null
        500:
            description: erro na criação do arquivo csv.
            content:
              application/json:
                schema:
                  $ref: "#components/schemas/ResponseMessage"
                example:
                  clientMessage: erro ao criar o arquivo csv.
                  serverMessage: null
                  data: null                        
  /athletes/{id}:
    put:
      tags:
        - Athletes
      summary: Atualiza informações de atleta.
      description: Atualiza informações de persona relacionada a atleta.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id(UUID) do atleta.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#components/schemas/Athlete"
            example:
              firstName: Mel
              lastName: Rodrigues
              birthDate: 2004-04-09
              phone: 11910434007
              picture: null
              category: sub-20
              isStarting: true
      responses:
        200:
          description: Senha de usuário atualizada.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/ResponseMessage"
              example:
                clientMessage: Treinador atualizado com sucesso.
                serverMessage: null
                data: null
        500:
          description: Erro genérico de servidor
          content:
            application/json:
              schema:
                $ref: "#components/schemas/ResponseMessage"
              example:
                clientMessage: Erro ao atualizar atleta.
                serverMessage: error message
                data: null
  /athletes/register-team/{id}:
    patch:
      tags:
        - Athletes
      summary: Registra o atleta em um time.
      description: Registra o atleta em um time.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id(UUID) do time.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
            example:
              id: ""
      responses:
        404:
          description: Atleta não encontrado.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/ResponseMessage"
              example:
                clientMessage: "Atleta não encontrado com o id: {id}."
                serverMessage: null
                data: null
        200:
          description: Atleta cadastrado no time.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/ResponseMessage"
              example:
                clientMessage: Atleta {lastName} cadatrado no time.
                serverMessage: null
                data: null
...